realtime drawing

+++ src/components/Editor/Pointers/Pointers.tsx
@@ -1,7 +1,8 @@
-import { Fragment, useEffect } from 'react';
+import { Fragment, useEffect, useRef } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { accountActions, AccountDataType } from '~/redux/stores';
-import { channel, usersChannel } from '~/supabase/config';
+import { FilesStateType } from '~/redux/filesSlice';
+import { accountActions, AccountDataType, ReducersType } from '~/redux/stores';
+import { realtimeUser } from '~/supabase/config';
 import PointerItem from './PointerItem';
 
 export type PointerProps = {
@@ -10,6 +11,9 @@ export type PointerProps = {
 
 function Pointers({ canvas }: PointerProps) {
   const accountData = useSelector<any, AccountDataType>((state) => state.account);
+  const { activeFile, isFilesLoading } = useSelector<ReducersType, FilesStateType>(
+    (state) => state.files,
+  );
   const dispatch = useDispatch();
 
   useEffect(() => {
@@ -17,39 +21,55 @@ function Pointers({ canvas }: PointerProps) {
       const { e } = event;
       const rect = (e.target as any)?.getBoundingClientRect();
 
-      await channel.send({
-        type: 'broadcast',
-        event: `location(${accountData.tempId})`,
-        payload: {
-          x: e.clientX - rect.left,
-          y: e.clientY - rect.top,
-          width: canvas?.width || 0,
-          height: canvas?.height || 0,
-        },
+      // await channel.send({
+      //   type: `broadcast`,
+      //   event: `location(${activeFile})(${accountData.tempId})`,
+      //   payload: {
+      //     x: e.clientX - rect.left,
+      //     y: e.clientY - rect.top,
+      //     width: canvas?.width || 0,
+      //     height: canvas?.height || 0,
+      //   },
+      // });
+
+      await realtimeUser.cursorChange({
+        activeFile,
+        tempAccountId: accountData.tempId,
+        x: e.clientX - rect.left,
+        y: e.clientY - rect.top,
+        width: canvas?.width || 0,
+        height: canvas?.height || 0,
       });
     }
 
     canvas?.on('mouse:move', handleMove);
     canvas?.on('touch:drag', handleMove);
-  }, [canvas]);
+
+    return () => {
+      canvas?.off('mouse:move', handleMove);
+      canvas?.off('touch:drag', handleMove);
+    };
+  }, [canvas, activeFile]);
 
   useEffect(() => {
-    channel.subscribe();
-    usersChannel
-      .on('presence', { event: 'sync' }, () => {
-        console.log('currently online users', usersChannel.presenceState());
-      })
-      .on('presence', { event: 'join' }, ({ newPresences }: any) => {
-        console.log('a new user has joined', newPresences);
-        dispatch(accountActions.addActiveUser(newPresences));
-      })
-      .on('presence', { event: 'leave' }, ({ leftPresences }: any) => {
-        console.log('a user has left', leftPresences);
-        dispatch(accountActions.removeActiveUser(leftPresences));
-      });
-  }, []);
+    if (!isFilesLoading && activeFile !== 'local') {
+      const usersChannel = realtimeUser.getUserChannel(activeFile);
+      usersChannel
+        .on('presence', { event: 'sync' }, () => {
+          console.log('currently online users', usersChannel.presenceState());
+        })
+        .on('presence', { event: 'join' }, ({ newPresences }: any) => {
+          console.log('a new user has joined', newPresences);
+          dispatch(accountActions.addActiveUser(newPresences));
+        })
+        .on('presence', { event: 'leave' }, ({ leftPresences }: any) => {
+          console.log('a user has left', leftPresences);
+          dispatch(accountActions.removeActiveUser(leftPresences));
+        });
+    }
+  }, [activeFile, isFilesLoading]);
 
-  console.log(accountData);
+  if (!canvas) return null;
 
   return (
     <Fragment>
